mina: much simplified

+++ lib/mina/org/apache/mina/core/polling/AbstractPollingIoConnector.java
@@ -44,25 +44,20 @@ import org.apache.mina.transport.socket.nio.NioSocketConnector;
 import org.apache.mina.util.ExceptionMonitor;
 
 /**
- * A base class for implementing client transport using a polling strategy. The
- * underlying sockets will be checked in an active loop and woke up when an
- * socket needed to be processed. This class handle the logic behind binding,
- * connecting and disposing the client sockets. A {@link Executor} will be used
- * for running client connection, and an {@link AbstractPollingIoProcessor} will
- * be used for processing connected client I/O operations like reading, writing
- * and closing.
+ * A base class for implementing client transport using a polling strategy.
+ * The underlying sockets will be checked in an active loop and woke up when an socket needed to be processed.
+ * This class handle the logic behind binding, connecting and disposing the client sockets.
+ * A {@link Executor} will be used for running client connection, and an {@link AbstractPollingIoProcessor}
+ * will be used for processing connected client I/O operations like reading, writing and closing.
  *
- * All the low level methods for binding, connecting, closing need to be
- * provided by the subclassing implementation.
+ * All the low level methods for binding, connecting, closing need to be provided by the subclassing implementation.
  *
  * @see NioSocketConnector for a example of implementation
  *
  * @author <a href="http://mina.apache.org">Apache MINA Project</a>
  */
 public abstract class AbstractPollingIoConnector extends AbstractIoConnector {
-
 	private final Queue<ConnectionRequest> connectQueue = new ConcurrentLinkedQueue<>();
-
 	private final Queue<ConnectionRequest> cancelQueue = new ConcurrentLinkedQueue<>();
 
 	private final IoProcessor<NioSession> processor;
@@ -75,10 +70,9 @@ public abstract class AbstractPollingIoConnector extends AbstractIoConnector {
 	private volatile boolean selectable;
 
 	/**
-	 * Constructor for {@link AbstractPollingIoConnector}. You need to provide a
-	 * default session configuration, a class of {@link IoProcessor} which will
-	 * be instantiated in a {@link SimpleIoProcessorPool} for better scaling in
-	 * multiprocessor systems. The default pool size will be used.
+	 * Constructor for {@link AbstractPollingIoConnector}. You need to provide a default session configuration,
+	 * a class of {@link IoProcessor} which will be instantiated in a {@link SimpleIoProcessorPool}
+	 * for better scaling in multiprocessor systems. The default pool size will be used.
 	 *
 	 * @see SimpleIoProcessorPool
 	 */
@@ -87,10 +81,9 @@ public abstract class AbstractPollingIoConnector extends AbstractIoConnector {
 	}
 
 	/**
-	 * Constructor for {@link AbstractPollingIoConnector}. You need to provide a
-	 * default session configuration, a class of {@link IoProcessor} which will
-	 * be instantiated in a {@link SimpleIoProcessorPool} for using multiple
-	 * thread for better scaling in multiprocessor systems.
+	 * Constructor for {@link AbstractPollingIoConnector}. You need to provide a default session configuration,
+	 * a class of {@link IoProcessor} which will be instantiated in a {@link SimpleIoProcessorPool}
+	 * for using multiple thread for better scaling in multiprocessor systems.
 	 *
 	 * @see SimpleIoProcessorPool
 	 *
@@ -101,17 +94,14 @@ public abstract class AbstractPollingIoConnector extends AbstractIoConnector {
 	}
 
 	/**
-	 * Constructor for {@link AbstractPollingIoAcceptor}. You need to provide a
-	 * default session configuration and an {@link Executor} for handling I/O
-	 * events. If null {@link Executor} is provided, a default one will be
-	 * created using {@link Executors#newCachedThreadPool()}.
+	 * Constructor for {@link AbstractPollingIoAcceptor}.
+	 * You need to provide a default session configuration and an {@link Executor} for handling I/O events.
+	 * If null {@link Executor} is provided, a default one will be created using {@link Executors#newCachedThreadPool()}.
 	 *
 	 * @see AbstractIoService#AbstractIoService(Executor)
 	 *
-	 * @param processor
-	 *            the {@link IoProcessor} for processing the {@link IoSession}
-	 *            of this transport, triggering events to the bound
-	 *            {@link IoHandler} and processing the chains of {@link IoFilter}
+	 * @param processor the {@link IoProcessor} for processing the {@link IoSession} of this transport,
+	 *            triggering events to the bound {@link IoHandler} and processing the chains of {@link IoFilter}
 	 */
 	private AbstractPollingIoConnector(IoProcessor<NioSession> processor) {
 		this.processor = processor;
@@ -142,8 +132,7 @@ public abstract class AbstractPollingIoConnector extends AbstractIoConnector {
 	protected abstract void init() throws IOException;
 
 	/**
-	 * Destroy the polling system, will be called when this {@link IoConnector}
-	 * implementation will be disposed.
+	 * Destroy the polling system, will be called when this {@link IoConnector} implementation will be disposed.
 	 */
 	protected abstract void destroy() throws IOException;
 
@@ -156,22 +145,20 @@ public abstract class AbstractPollingIoConnector extends AbstractIoConnector {
 	protected abstract SocketChannel newHandle(SocketAddress localAddress) throws IOException;
 
 	/**
-	 * Connect a newly created client socket handle to a remote
-	 * {@link SocketAddress}. This operation is non-blocking, so at end of the
-	 * call the socket can be still in connection process.
+	 * Connect a newly created client socket handle to a remote {@link SocketAddress}.
+	 * This operation is non-blocking, so at end of the call the socket can be still in connection process.
 	 *
 	 * @param channel the client socket handle
 	 * @param remoteAddress the remote address where to connect
-	 * @return <tt>true</tt> if a connection was established, <tt>false</tt> if
-	 *         this client socket is in non-blocking mode and the connection operation is in progress
+	 * @return <tt>true</tt> if a connection was established,
+	 *         <tt>false</tt> if this client socket is in non-blocking mode and the connection operation is in progress
 	 * @throws IOException If the connect failed
 	 */
 	protected abstract boolean connect(SocketChannel channel, SocketAddress remoteAddress) throws IOException;
 
 	/**
-	 * Finish the connection process of a client socket after it was marked as
-	 * ready to process by the {@link #select(int)} call. The socket will be
-	 * connected or reported as connection failed.
+	 * Finish the connection process of a client socket after it was marked as ready to process
+	 * by the {@link #select(int)} call. The socket will be connected or reported as connection failed.
 	 *
 	 * @param channel the client socket handle to finish to connect
 	 * @return true if the socket is connected
@@ -180,8 +167,7 @@ public abstract class AbstractPollingIoConnector extends AbstractIoConnector {
 
 	/**
 	 * Create a new {@link IoSession} from a connected socket client handle.
-	 * Will assign the created {@link IoSession} to the given
-	 * {@link IoProcessor} for managing future I/O events.
+	 * Will assign the created {@link IoSession} to the given {@link IoProcessor} for managing future I/O events.
 	 *
 	 * @param processor1 the processor in charge of this session
 	 * @param channel the newly connected client socket handle
@@ -202,9 +188,8 @@ public abstract class AbstractPollingIoConnector extends AbstractIoConnector {
 	protected abstract void wakeup();
 
 	/**
-	 * Check for connected sockets, interrupt when at least a connection is
-	 * processed (connected or failed to connect). All the client socket
-	 * descriptors processed need to be returned by {@link #selectedHandles()}
+	 * Check for connected sockets, interrupt when at least a connection is processed (connected or failed to connect).
+	 * All the client socket descriptors processed need to be returned by {@link #selectedHandles()}
 	 *
 	 * @param timeout The timeout for the select() method
 	 * @return The number of socket having received some data
@@ -311,7 +296,7 @@ public abstract class AbstractPollingIoConnector extends AbstractIoConnector {
 			while (selectable) {
 				try {
 					// the timeout for select shall be smaller of the connect timeout or 1 second...
-					int selected = select(Math.min(getConnectTimeoutMillis(), 1000L));
+					int selected = select(Math.min(getConnectTimeoutMillis(), 1000));
 
 					nHandles += registerNew();
 
@@ -407,8 +392,7 @@ public abstract class AbstractPollingIoConnector extends AbstractIoConnector {
 		}
 
 		/**
-		 * Process the incoming connections, creating a new session for each valid
-		 * connection.
+		 * Process the incoming connections, creating a new session for each valid connection.
 		 */
 		private int processConnections(Iterator<SocketChannel> handlers) {
 			int nHandles = 0;
@@ -471,21 +455,12 @@ public abstract class AbstractPollingIoConnector extends AbstractIoConnector {
 		/** The time up to this connection request will be valid */
 		private final long deadline;
 
-		/**
-		 * Creates a new ConnectionRequest instance
-		 *
-		 * @param channel The IoHander
-		 * @param callback The IoFuture callback
-		 */
 		public ConnectionRequest(SocketChannel channel) {
 			this.channel = channel;
-			long timeout = getConnectTimeoutMillis();
+			int timeout = getConnectTimeoutMillis();
 			deadline = (timeout > 0L ? System.currentTimeMillis() + timeout : Long.MAX_VALUE);
 		}
 
-		/**
-		 * @return The IoHandler instance
-		 */
 		public SocketChannel getHandle() {
 			return channel;
 		}
